Agrego mas tiempo de espera para que el servidor pueda responder. Tomo valores puerto y nombre de red desde el archivo de configuracion.

+++ validar-echo-server.sh
@@ -2,6 +2,10 @@
 
 # Obtener el nombre de la red desde docker-compose
 NOMBRE_RED=$(docker network ls --format "{{.Name}}" | grep testing_net)
+SERVER_NAME="server"
+SERVER_PORT=$(docker exec "$SERVER_NAME" sh -c "awk -F '=' '/SERVER_PORT/ {print \$2}' /config.ini | tr -d ' '")
+MENSAJE="TestMessage123"
+TIEMPO_ESPERA=100
 
 # Verificar si la red existe
 if [ -z "$NOMBRE_RED" ]; then
@@ -9,17 +13,18 @@ if [ -z "$NOMBRE_RED" ]; then
     exit 1
 fi
 
-SERVER_NAME="server"
-SERVER_PORT=12345
-MENSAJE="TestMessage123"
-TIEMPO_ESPERA=2
+# Verificar si el puerto existe
+if [ -z "$SERVER_PORT" ]; then
+    echo "Error: El puerto del servidor no está disponible."
+    exit 1
+fi
 
 # Ejecutar un contenedor temporal en la red de Docker para probar la conexión con netcat
 RESPUESTA=$(docker run --rm --network "$NOMBRE_RED" busybox sh -c \
     "echo '$MENSAJE' | nc -w $TIEMPO_ESPERA $SERVER_NAME $SERVER_PORT")
 
 # Verificar si la respuesta es igual al mensaje enviado
-if [ "$RESPUESTA" == "$MENSAJE" ]; then
+if [ "$RESPUESTA" = "$MENSAJE" ]; then
     echo "action: test_echo_server | result: success"
     exit 0
 else
